pyproject.toml
[project]
name = "dcag"
version = "0.1.0"
description = "Data Context Abstraction Graph — competence-aware workflow engine"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1",
    "jsonschema>=4.20",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
